helm changes completed for app and db with dynamic inputs
