+++ archival-older-scriptsforRemoval/app-deploy.yaml
@@ -1,0 +1,30 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: restapi
+  labels:
+    app: restapi
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: restapi
+  template:
+    metadata:
+      labels:
+        app: restapi
+    spec:
+      containers:
+        - name: api
+          image: althaf/pprorepo:22
+          imagePullPolicy: IfNotPresent
+          ports:
+            - containerPort: 8000
+          volumeMounts:
+            - name: config-volume
+              mountPath: /usr/app/app-config.yaml
+              subPath: app-config.yaml 
+      volumes:
+        - name: config-volume
+          configMap:
+            name: restapi-cf
