+++ archival-older-scriptsforRemoval/app-svc.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: restapi-service
+spec:
+  type: NodePort
+  selector:
+    app: restapi
+  ports:
+    - protocol: TCP
+      port: 8000
+      targetPort: 8000
