+++ archival-older-scriptsforRemoval/pg-config.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: pg-secret
+type: Opaque
+data:
+  pg-password: cGFzcwo=
