--- devops/code-deploy/configs/pproapp/dev/dev-app.yaml
@@ -1,4 +1,0 @@
-meta:
-  pproapp:
-    database:
-      password: 'cGFzcwo='
