--- devops/code-deploy/configs/pproapp/prod/prod-infra.yaml
@@ -1,4 +1,0 @@
-meta:
-  pproapp:
-    database:
-      password: 'cGFzcw2341@='
