--- devops/code-deploy/helm/app/app-deploy.yaml
@@ -1,28 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: restapi
-  labels:
-    app: restapi
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: restapi
-  template:
-    metadata:
-      labels:
-        app: restapi
-    spec:
-      containers:
-      - name: api
-        image: althaf/pprorepo:9
-        imagePullPolicy: IfNotPresent
-        ports:
-          - containerPort: 8000
-        env:
-            - name: POSTGRES_PASSWORD
-              valueFrom:
-                secretKeyRef:
-                  name: pg-secret
-                  key: pg-password
