--- devops/code-deploy/helm/app/app-svc.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: restapi-service
-spec:
-  type: NodePort
-  selector:
-    app: restapi
-  ports:
-    - protocol: TCP
-      port: 8000
-      targetPort: 8000
