--- devops/code-deploy/helm/db/pg-config.yaml
@@ -1,7 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: pg-secret
-type: Opaque
-data:
-  pg-password: cGFzcwo=
