--- devops/code-deploy/helm/db/pg-svc.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: pg-service
-spec:
-  selector:
-    app: pg
-  ports:
-    - protocol: TCP
-      port: 5432
-      targetPort: 5432
