+++ devops/deploy/configs/ppro-sample-service/prod/app-config.yaml
@@ -1,0 +1,4 @@
+DB_USER: "postgres"
+DB_PASS: "prodpass"
+DB_NAME: "postgres"
+DB_ENDPOINT: "pg-service"
