+++ devops/deploy/helm/app/appchart/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "appchart.fullname" . }}-test-connection"
+  labels:
+    {{- include "appchart.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args: ['{{ include "appchart.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
