+++ devops/deploy/helm/db/dbchart/templates/configmap.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: pg-secret
+type: Opaque
+data:
+  pg-password: {{ .Values.dbPass }}
