+++ devops/deploy/helm/db/dbchart/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "dbchart.fullname" . }}-test-connection"
+  labels:
+    {{- include "dbchart.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args: ['{{ include "dbchart.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
