+++ ppro-sample-service/app/scripts/init.sql
@@ -1,0 +1,7 @@
+
+CREATE TABLE IF NOT EXISTS pprotble(
+    id VARCHAR,
+    val VARCHAR
+);
+
+INSERT INTO pprotble(id, val) VALUES (1, 'Hello World!');
